+++ protolink/models/__init__.py
@@ -2,7 +2,7 @@ from protolink.client.request_spec import ClientRequestSpec
 from protolink.core.agent_card import AgentCard, AgentSkill
 from protolink.core.artifact import Artifact
 from protolink.core.message import Message
-from protolink.core.part import Part
+from protolink.core.part import Part, RouteDecision
 from protolink.core.task import Task, TaskState
 from protolink.server.endpoint_handler import EndpointSpec
 
@@ -14,6 +14,7 @@ __all__ = [
     "EndpointSpec",
     "Message",
     "Part",
+    "RouteDecision",
     "Task",
     "TaskState",
 ]